+++ connectors/hazelcast/hazelcast-connector/src/main/java/org/platformlambda/hazelcast/reporter/PresenceManager.java
@@ -216,5 +216,4 @@ public class PresenceManager extends Thread {
         normal = false;
     }
 
-
 }